requirements.txt
openai>=1.14.0
instructor>=1.0.0
pydantic>=2.6.0
tenacity>=8.2.2
python-dotenv>=1.0.1